- fix field int to int64 - fix corrupt problem in building application

+++ api/daos/gorms/ArbitrosGorm.go
@@ -1,8 +1,8 @@
 package gorms
 
 type ArbitrosGorm struct {
-	IDArbitro int `gorm:"column:id_arbitro;primary_key"`
-	IDPersona int `gorm:"column:id_persona"`
+	IDArbitro int64 `gorm:"column:id_arbitro;primary_key"`
+	IDPersona int64 `gorm:"column:id_persona"`
 }
 
 // TableName sets the insert table name for this struct type
