+++ api/daos/gorms/CampeonatosGorm.go
@@ -9,8 +9,8 @@ type CampeonatosGorm struct {
 	Descripcion  sql.NullString `gorm:"column:descripcion"`
 	FechaFin     time.Time      `gorm:"column:fecha_fin"`
 	FechaInicio  time.Time      `gorm:"column:fecha_inicio"`
-	IDCampeonato int            `gorm:"column:id_campeonato;primary_key"`
-	IDLiga       int            `gorm:"column:id_liga"`
+	IDCampeonato int64            `gorm:"column:id_campeonato;primary_key"`
+	IDLiga       int64            `gorm:"column:id_liga"`
 	IDModelo     sql.NullString `gorm:"column:id_modelo"`
 }
