+++ api/daos/gorms/EquiposJugadoresGorm.go
@@ -1,8 +1,8 @@
 package gorms
 
 type EquiposJugadoresGorm struct {
-	IDEquipos   int `gorm:"column:id_equipos;primary_key"`
-	IDJugadores int `gorm:"column:id_jugadores;primary_key"`
+	IDEquipos   int64 `gorm:"column:id_equipos;primary_key"`
+	IDJugadores int64 `gorm:"column:id_jugadores;primary_key"`
 }
 
 // TableName sets the insert table name for this struct type
