+++ api/daos/gorms/SancionesEquiposGorm.go
@@ -1,8 +1,8 @@
 package gorms
 
 type SancionesEquiposGorm struct {
-	IDEquipo    int `gorm:"column:id_equipo;primary_key"`
-	IDSanciones int `gorm:"column:id_sanciones;primary_key"`
+	IDEquipo    int64 `gorm:"column:id_equipo;primary_key"`
+	IDSanciones int64 `gorm:"column:id_sanciones;primary_key"`
 }
 
 // TableName sets the insert table name for this struct type
