+++ api/daos/gorms/SancionesGorm.go
@@ -4,8 +4,8 @@ import "database/sql"
 
 type SancionesGorm struct {
 	Descripcion   sql.NullString `gorm:"column:descripcion"`
-	IDLigas       int            `gorm:"column:id_ligas"`
-	IDSanciones   int            `gorm:"column:id_sanciones;primary_key"`
+	IDLigas       int64            `gorm:"column:id_ligas"`
+	IDSanciones   int64            `gorm:"column:id_sanciones;primary_key"`
 	Observaciones sql.NullString `gorm:"column:observaciones"`
 }
